+++ livedata/src/main/java/io/getstream/chat/android/livedata/ChatDomain.kt
@@ -96,6 +96,7 @@ interface ChatDomain {
 
         private var userPresence: Boolean = false
         private var offlineEnabled: Boolean = true
+        private var backgroundSyncEnabled: Boolean = true
         private var recoveryEnabled: Boolean = true
         private var backgroundSyncConfig: BackgroundSyncConfig = BackgroundSyncConfig.UNAVAILABLE
         private var notificationConfig: NotificationConfig = NotificationConfigUnavailable
@@ -108,10 +109,17 @@ interface ChatDomain {
 
         fun backgroundSyncEnabled(apiKey: String, userToken: String): Builder {
             // TODO: Consider exposing apiKey and userToken by ChatClient to make this function more friendly
-            this.backgroundSyncConfig = BackgroundSyncConfig(apiKey, user.id, userToken)
             if (apiKey.isEmpty() || user.id.isEmpty() || userToken.isEmpty()) {
-                throw IllegalArgumentException("ChatDomain.Builder::backgroundSyncEnabled. apiKey and user.id, and userToken must not be empty.")
+                this.backgroundSyncConfig = BackgroundSyncConfig.UNAVAILABLE
+            } else {
+                this.backgroundSyncConfig = BackgroundSyncConfig(apiKey, user.id, userToken)
             }
+            return this
+        }
+
+        fun backgroundSyncDisabled(): Builder {
+            this.backgroundSyncEnabled = false
+            this.backgroundSyncConfig = BackgroundSyncConfig.UNAVAILABLE
             return this
         }
 
@@ -151,8 +159,14 @@ interface ChatDomain {
         }
 
         fun build(): ChatDomain {
-            storeBackgroundSyncConfig(backgroundSyncConfig)
-            storeNotificationConfig(notificationConfig)
+            if (backgroundSyncEnabled) {
+                if (backgroundSyncConfig == BackgroundSyncConfig.UNAVAILABLE) {
+                    throw IllegalStateException("ChatDomain.Builder::build. backgroundSyncEnabled must be called with non-empty params.")
+                } else {
+                    storeBackgroundSyncConfig(backgroundSyncConfig)
+                    storeNotificationConfig(notificationConfig)
+                }
+            }
             instance = buildImpl()
             return instance
         }