+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 # To be released:
+- Improve sync data validation in ChatDomain.Builder
 
 - Significant performance improvements to offline storage
 - Default message limit for the queryChannels use case changed from 10 to 1. This is a more sensible default for the channel list view of most chat apps